requirements.txt
absl-py==0.9.0
alabaster==0.7.12
apex==0.1
appdirs==1.4.3
ascii-graph==1.5.1
asn1crypto==1.3.0
attrs==19.3.0
audioread==2.1.8
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.3
boto3==1.12.28
botocore==1.15.28
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
codecov==2.0.22
conda==4.8.2
conda-build==3.18.11
conda-package-handling==1.6.0
configparser==5.0.1
coverage==5.0.4
cryptography==2.8
cycler==0.10.0
cymem==2.0.2
Cython==0.28.4
cytoolz==0.9.0.1
DataProperty==0.46.4
decorator==4.4.1
defusedxml==0.6.0
dill==0.2.9
docker-pycreds==0.4.0
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
glob2==0.7
google-auth==1.11.3
google-auth-oauthlib==0.4.1
grpcio==1.27.2
h5py==2.10.0
html2text==2020.1.16
hypothesis==4.50.8
idna==2.8
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.5.0
inflect==4.1.0
ipdb==0.13.2
ipykernel==5.2.0
ipython==7.12.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
json5==0.9.3
jsonschema==3.2.0
jupyter-client==6.1.0
jupyter-core==4.6.3
jupyter-tensorboard==0.1.10
jupyterlab==1.0.4
jupyterlab-server==1.0.7
jupytext==1.4.1
kiwisolver==1.1.0
libarchive-c==2.8
librosa==0.6.3
lief==0.9.0
llvmlite==0.28.0
lmdb==0.98
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
maskrcnn-benchmark==0.1
matplotlib==3.2.1
mbstrdecoder==0.8.4
mccabe==0.6.1
mistune==0.8.4
mlperf-compliance==0.0.10
mock==4.0.1
more-itertools==8.2.0
msgfy==0.0.7
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.0
nltk==3.4.5
notebook==6.0.3
numba==0.43.1
numpy==1.18.1
nvidia-dali==0.19.0
oauthlib==3.1.0
onnx==1.6.0
opencv-python==3.4.1.15
packaging==20.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.6.1
pathtools==0.1.2
pathvalidate==2.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow-SIMD==5.3.0.post1
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
progressbar==2.5
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.3
protobuf==3.14.0
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycocotools==2.0+nv0.4.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycuda==2019.1.2
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
pytablewriter==0.50.0
pytest==5.4.1
pytest-cov==2.8.1
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
python-hostlist==1.20
python-nvd3==0.15.0
python-slugify==4.0.0
pytools==2020.1
pytorch-transformers==1.1.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==19.0.0
regex==2018.1.10
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
revtok==0.0.3
rsa==4.0
ruamel-yaml==0.15.87
s3transfer==0.3.3
sacrebleu==1.2.10
sacremoses==0.0.35
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.85
sentry-sdk==0.19.5
shortuuid==1.0.1
six==1.14.0
smmap==3.0.5
snowballstemmer==2.0.0
SoundFile==0.10.3.post1
soupsieve==1.9.5
sox==1.3.7
spacy==2.0.16
Sphinx==2.4.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SSD==0.1
subprocess32==3.5.4
subword-nmt==0.3.3
tabledata==1.1.0
tabulate==0.8.7
tensorboard==2.1.0
tensorrt==7.0.0.11
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thinc==6.12.1
toml==0.10.0
toolz==0.10.0
torch==1.5.0a0+8f84ded
torchtext==0.4.0
torchvision==0.6.0a0
tornado==6.0.4
tqdm==4.31.1
traitlets==4.3.3
typepy==0.6.6
typing==3.7.4.1
typing-extensions==3.7.4.1
ujson==1.35
Unidecode==1.1.1
urllib3==1.25.8
wandb==0.10.15
watchdog==0.10.4
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.10.11
yacs==0.1.6
zipp==3.1.0
